+++ CHANGELOG.md
@@ -5,12 +5,15 @@
 #### Breaking Changes
 
 * [QAMenu] Renamed parameter `isSearchable` to `isPaneSearchable` in `Item.asChildPaneItem()` (MR #6)
+* [QAMenu] Removed parameter `trigger` from `QAMenu()` in favor of `QAMenu.setTrigger()` (MR #7)
+* [QAMenu] Renamed `QAMenuDismissBehavior` to `QAMenu.DismissBehavior()` (MR #7)
 
 #### Enhancements
 
 * [QAMenu] Added missing `asChildPaneItem()` options to create items with values (MR #6)
 * [QAMenu] Allow to pass the root pane after initializing `QAMenu` with `setRootPane()` (MR #7)
-
+* [QAMenu] `QAMenu.Trigger` configuration is now backed by NSUserDefaults  (MR #7)
+* [QAMenu] Exposed configuration option `QAMenu.DismissBehavior` (MR #7)
 
 #### Bug Fixes